+++ models/sections.go
@@ -5,14 +5,14 @@ import (
 )
 
 type Section struct {
-	SectionID int64    `gorm:"column:section_id; primaryKey;autoIncrement" json:"section_id"`
+	SectionID int      `gorm:"column:section_id; primaryKey;autoIncrement" json:"section_id"`
 	Name      string   `gorm:"column:name; not null; type:VARCHAR(255);" json:"name" binding:"required,max=255"`
 	ImageUrl  string   `gorm:"column:image_url; default:NULL; type:LONGTEXT;" json:"image_url" binding:"max=5000"`
 	Groups    []Groups `gorm:"many2many:sections_groups;foreignKey:SectionID;joinForeignKey:SectionID;References:GroupID;joinReferences:GroupID" json:"-"`
 }
 
 type Groups struct {
-	GroupID int64  `gorm:"column:group_id;primaryKey;autoIncrement;" json:"group_id"`
+	GroupID int    `gorm:"column:group_id;primaryKey;autoIncrement;" json:"group_id"`
 	Name    string `gorm:"column:name; not null; type:VARCHAR(255)" json:"name" binding:"required,max=255"`
 }
 
@@ -25,7 +25,7 @@ func (Section) TableName() string {
 // }
 
 type UpdateSectionModels struct {
-	SectionID int64  `gorm:"column:section_id; primaryKey;autoIncrement" json:"section_id"`
+	SectionID int    `gorm:"column:section_id; primaryKey;autoIncrement" json:"section_id"`
 	Name      string `gorm:"column:name; not null; type:VARCHAR(255);" json:"name" binding:"max=255"`
 	ImageUrl  string `gorm:"column:image_url; default:NULL; type:LONGTEXT;" json:"image_url" binding:"max=5000"`
 }
